добавил отображение количества и суммы заказа при повторном вводе

+++ ruspovar_scala/src/main/scala/MenuItem.scala
@@ -2,6 +2,6 @@
 
 sealed abstract class MenuItem(val line: String)
 
-final case class EdibleMenuItem(override val line: String, subTitle: Option[String], var quantity: Int, price: BigDecimal) extends MenuItem(line)
+final case class EdibleMenuItem(override val line: String, subTitle: Option[String], var quantity: Int, price: Int) extends MenuItem(line)
 
 final case class SubTitleMenuItem(override val line: String) extends MenuItem(line)
